logrotate logfile on EL6 - fixes #14 & #15
consul logs by default to STDOUT which is captured by systemd on
newer systems. On a sysv-init based system - like EL6 - we redirect
that to our own logfile. Which then just grows continously. As
enabling logging to syslog in consul does not silence output
to stdout, we would now log twice. So on these systems we should
logrotate that logfile to avoid growing the file endless.