consul.spec: Make version outside controllable
By using RPM macros as, or similar to, the provided, it becomes possible
to simply specify the 'hard-coded' version in the SPEC file itself, or,
if one so chooses, by e.g. 'rpmbuild --define "_version 0.6.3"'.

This removes a piece of state from the spec file which can be handy,
if one uses consul-rpm to build but wants to avoid patching the spec
file on every build, to build another version.